requirements.txt
MySQL-python
django-bootstrap3